requirements.txt
django==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
redgreenunittest==0.1.1
drf-spectacular==0.25.1
django-cors-headers==3.13.0
django-storages==1.13.1
pillow==9.3.0
boto3==1.26.14
botocore~=1.29.27
s3transfer~=0.6.0
setuptools~=65.5.1
python-dotenv~=0.21.0
asgiref~=3.5.2
sqlparse~=0.4.3
python-dateutil~=2.8.2
urllib3~=1.26.13
django-filter~=22.1
mysqlclient==2.1.1
django-mysql==3.8.1